# swagger.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kasir Laundry Pro",
    "version": "1.0.0",
    "description": "API Routes and schema details of Kasir Laundry Pro Services"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
...
    "/user/upload-logo": {
      "post": {
        "summary": "Uploads a temporary logo file",
        "description": "Uploads a logo file to Supabase Storage and returns the file path.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The logo file to upload"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "filePath": {
                      "type": "string",
                      "description": "The path of the uploaded file in Supabase Storage"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Tidak ada file yang diunggah"
          },
          "500": {
            "description": "Gagal mengunggah file"
          }
        }
      }
    },
    "/user/details": {
      "get": {
        "summary": "Retrieve user details",
        "description": "Fetches the details of a user by their ID.",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "User details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User tidak ditemukan"
          }
        }
      }
    },
    "/user/{id}": {
      "put": {
        "summary": "Update user details and optionally submit logo",
        "description": "Updates user details including optional logo upload. If a logo file path is provided, it will be moved from temporary to final location.",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the user to update.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the user"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email address of the user"
                  },
                  "phone_number": {
                    "type": "string",
                    "description": "Phone number of the user"
                  },
                  "address": {
                    "type": "string",
                    "description": "Address of the user"
                  },
                  "logo": {
                    "type": "string",
                    "description": "Path of the uploaded logo in Supabase Storage"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rincian pengguna berhasil diperbarui"
          },
          "400": {
            "description": "Missing required fields or invalid input"
          },
          "404": {
            "description": "Pengguna tidak ditemukan"
          },
          "500": {
            "description": "Failed to update user details or process logo"
          }
        }
      }
    },
    "/user/delete-temp-files": {
      "get": {
        "summary": "Delete temporary files",
        "description": "Deletes all files in the temporary storage folder. Requires a valid secret token in the request header.",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "cron-job-token",
            "required": true,
            "description": "The secret token required to authorize the request.",
            "schema": {
              "type": "string",
              "example": "d5f811"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Files deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Files already deleted"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Invalid token"
          },
          "500": {
            "description": "Error deleting files"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User's unique identifier"
          },
          "name": {
            "type": "string",
            "description": "Name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "john.doe@example.com"
          },
          "oauth": {
            "type": "boolean",
            "description": "Whether the user is authenticated via OAuth"
          },
          "token": {
            "type": "string",
            "description": "Authentication token for the user"
          },
          "phone_number": {
            "type": "string",
            "description": "Phone number of the user (nullable)",
            "nullable": true,
            "example": "+12345678901"
          },
          "logo": {
            "type": "string",
            "description": "URL of the user's logo (nullable)",
            "nullable": true,
            "example": "https://example.com/logo.png"
          },
          "address": {
            "type": "string",
            "description": "Address of the user (nullable)",
            "nullable": true,
            "example": "123 Main St, Springfield"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of when the user was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of when the user was last updated"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
...
    }
  },
...
}